Refine tests, further fixes to PC and JR
When a jr command happens, it was only inc PC if it actually jumped,
but this isn't how it should work, its a safety mechanism.
Always inc on every instruction that uses a parameter.

Current issue to note: Addr on jumps are always +1 off,
possibly need to increase current address. I guess I am
still assuming current address was a previous one..
Look to increment PC for this address jump.